Fixed output

+++ src/driver.cpp
@@ -45,7 +45,7 @@ bool testTemperatureSensor() {
   cout << "Initialized Temperature Sensor" << endl;
   cout << "Reading Temperature data..." << endl;
   for (int i = 0; i < 5; i++) {
-    cout << "Temperature = " << tempSensor.readTemp() << endl;
+    cout << "Temperature = " << (int) tempSensor.readTemp() << endl;
     usleep(500000);
   }
   return true;
